Refactor cart summary schema and calculations for JSON compatibility
- Changed the CartSummary fields from Decimal to float to ensure proper JSON serialization.
- Updated the tax calculation logic to apply GST on the subtotal directly, reflecting the correct total calculation.
- Converted all monetary values to float in the cart summary response for consistent data handling.

+++ app/schemas/cart.py
@@ -40,11 +40,11 @@ class CartItemResponse(BaseModel):
 
 
 class CartSummary(BaseModel):
-    subtotal: Decimal
-    discount: Decimal
-    delivery_charge: Decimal
-    tax: Decimal
-    total: Decimal
+    subtotal: float  # Changed from Decimal to float for JSON serialization
+    discount: float
+    delivery_charge: float
+    tax: float
+    total: float
     item_count: int = 0
